Integration on more features is done.
