+++ src/test/java/com/azulCRM/step_definitions/AddLinkInMessage_Step_Defs.java
@@ -7,31 +7,18 @@ import io.cucumber.java.en.Then;
 import io.cucumber.java.en.When;
 import org.junit.Assert;
 
-public class AddLinkInMessageStepDefs {
-
-    // testing VALID URL
+public class AddLinkInMessage_Step_Defs {
     private ActivityStreamPage ACTIVITY;
+
     @Before
     public void setUpPages() {
         ACTIVITY = new ActivityStreamPage();
-    }
-
-      /*  @When("user clicks on activity stream button")
-    public void user_clicks_on_activity_stream_button() {
-        ActivityStreamPage.activityStreamBtn.click();
-    }
-       */
-
-    @When("user clicks on massage button")
-    public void user_clicks_on_massage_button() {
-        ACTIVITY.messageBtn.click();
     }
 
     @When("user clicks on link button")
     public void user_clicks_on_link_button() {
         ACTIVITY.addLink.click();
     }
-
 
     @When("user enters {string} in text box")
     public void user_enters_in_text_box(String linkText) {
@@ -43,7 +30,6 @@ public class AddLinkInMessageStepDefs {
         ACTIVITY.urlInput.sendKeys(linkUrl);
     }
 
-
     @When("user clicks on Save button")
     public void user_clicks_on_save_button() {
         BrowserUtils.sleep(1);
@@ -52,20 +38,7 @@ public class AddLinkInMessageStepDefs {
 
     @Then("user sees the {string} as a clickable link.")
     public void user_sees_the_as_a_clickable_link(String linkText) {
-
-        System.out.println("linkText = " + linkText);
      String expectedLinkText = "google search page";
-        System.out.println("expectedLinkText = " + expectedLinkText);
      Assert.assertTrue(linkText.contains(expectedLinkText));
-
-        BrowserUtils.sleep(2);
     }
-
-    @When("user clicks on send button")
-    public void user_clicks_on_send_button() {
-
-             ACTIVITY.sendBtn.click();
-    }
-
-
 }
